Adds icons to pull method
Adds icons to our list of things to pull in from the Sass Framework. This will allow our icons to live in our Pattern Library and get version controlled. This is also useful so that we can register our icon SVGs as partials to easily include within our handlebars templates.