Creates passthrough helpers for handlebars
This enables us to dynamically generate additional key-value pairs for our data on the fly, where existing data is passed through one or more handlebars helpers. The transformed value is then saved in to original data object underneath its new key. This is helpful for when we need to transform things like dates and times to other formats.